pyproject.toml
[project]
name = "brain-jepa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.40.64",
    "einops>=0.8.1",
    "gdown>=5.2.0",
    "ipython>=8.18.1",
    "nibabel>=5.3.2",
    "numpy>=2.0.2",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.13.1",
    "tensorboard>=2.20.0",
    "timm>=1.0.21",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
]
